only push to git if not released repo

+++ workflow-fedora-based-buildroot.sh
@@ -92,45 +92,46 @@ cat ${PACKAGELIST_DIR}/Packages.added.tmp | sort | uniq -cd | sed -n -e 's/^ *4 
 cat ${PACKAGELIST_DIR}/Packages.Binary.Buildroot.tmp | sort | uniq -cd | sed -n -e 's/^ *4 \(.*\)/\1/p' | sort -u -o ${PACKAGELIST_DIR}/Packages.Binary.Buildroot.common
 sort -u -o ${PACKAGELIST_DIR}/Packages.Source.Buildroot.NVR.all-arches ${PACKAGELIST_DIR}/Packages.Source.Buildroot.NVR.all-arches
 
-# Generate the buildroot workload
-rm -f ${PACKAGELIST_DIR}/${VIEW}-buildroot-workload.yaml
-cat ${WORK_DIR}/conf/${VIEW}-buildroot-workload.head >> ${PACKAGELIST_DIR}/${VIEW}-buildroot-workload.yaml
-#cat ${PACKAGELIST_DIR}/Packages.Binary.Buildroot.common | awk '{print "        - " $1}' >> ${PACKAGELIST_DIR}/${VIEW}-buildroot-workload.yaml
-cat ${PACKAGELIST_DIR}/Packages.added.common | awk '{print "        - " $1}' >> ${PACKAGELIST_DIR}/${VIEW}-buildroot-workload.yaml
-echo "    arch_packages:" >> ${PACKAGELIST_DIR}/${VIEW}-buildroot-workload.yaml
-for this_arch in ${ARCH_LIST[@]}
-do
-  DATA_DIR="${DATA_DIR_BASE}/${this_arch}"
-  echo "        ${this_arch}:" >> ${PACKAGELIST_DIR}/${VIEW}-buildroot-workload.yaml
-  comm -13 ${PACKAGELIST_DIR}/Packages.added.common ${DATA_DIR}/${NEW_DIR}/added-binary-package-names.txt | awk '{print "            - " $1}' >> ${PACKAGELIST_DIR}/${VIEW}-buildroot-workload.yaml
-  #comm -13 ${PACKAGELIST_DIR}/Packages.Binary.Buildroot.common ${DATA_DIR}/${NEW_DIR}/buildroot-binary-package-names.txt | awk '{print "            - " $1}' >> ${PACKAGELIST_DIR}/${VIEW}-buildroot-workload.yaml
-done
+if ! [ "${REPO_BASE}" == "released" ] ; then
+  # Generate the buildroot workload
+  rm -f ${PACKAGELIST_DIR}/${VIEW}-buildroot-workload.yaml
+  cat ${WORK_DIR}/conf/${VIEW}-buildroot-workload.head >> ${PACKAGELIST_DIR}/${VIEW}-buildroot-workload.yaml
+  cat ${PACKAGELIST_DIR}/Packages.added.common | awk '{print "        - " $1}' >> ${PACKAGELIST_DIR}/${VIEW}-buildroot-workload.yaml
+  echo "    arch_packages:" >> ${PACKAGELIST_DIR}/${VIEW}-buildroot-workload.yaml
+  for this_arch in ${ARCH_LIST[@]}
+  do
+    DATA_DIR="${DATA_DIR_BASE}/${this_arch}"
+    echo "        ${this_arch}:" >> ${PACKAGELIST_DIR}/${VIEW}-buildroot-workload.yaml
+    comm -13 ${PACKAGELIST_DIR}/Packages.added.common ${DATA_DIR}/${NEW_DIR}/added-binary-package-names.txt | awk '{print "            - " $1}' >> ${PACKAGELIST_DIR}/${VIEW}-buildroot-workload.yaml
+    #comm -13 ${PACKAGELIST_DIR}/Packages.Binary.Buildroot.common ${DATA_DIR}/${NEW_DIR}/buildroot-binary-package-names.txt | awk '{print "            - " $1}' >> ${PACKAGELIST_DIR}/${VIEW}-buildroot-workload.yaml
+  done
 
-# Trim buildroot workload of packages we don't want in there
-for pkgname in ${BAD_PACKAGES[@]}
-do
-  sed -i "/ ${pkgname}$/d" ${PACKAGELIST_DIR}/${VIEW}-buildroot-workload.yaml
-done
+  # Trim buildroot workload of packages we don't want in there
+  for pkgname in ${BAD_PACKAGES[@]}
+  do
+    sed -i "/ ${pkgname}$/d" ${PACKAGELIST_DIR}/${VIEW}-buildroot-workload.yaml
+  done
 
-# Upload to feedback-pipeline
-if ! [ -d ${GIT_DIR}/feedback-pipeline-config ] ; then
-  mkdir -p ${GIT_DIR}
-  cd ${GIT_DIR}
-  git clone git@github.com:minimization/feedback-pipeline-config.git
-fi
-if ! [ -d ${GIT_DIR}/feedback-pipeline-config ] ; then
-	echo
-	echo "You do not seem to have correct credentials for the git repo"
-	echo "Exiting so you do no harm"
-	echo
-	exit 5
-fi
-cd ${GIT_DIR}/feedback-pipeline-config/configs
-git pull
-cp ${PACKAGELIST_DIR}/${VIEW}-buildroot-workload.yaml .
-git add ${VIEW}-buildroot-workload.yaml
-git commit -m "Update ${VIEW}-buildroot-workload $(date +%Y-%m-%d-%H:%M)"
-git push
+  # Upload to feedback-pipeline
+  if ! [ -d ${GIT_DIR}/feedback-pipeline-config ] ; then
+    mkdir -p ${GIT_DIR}
+    cd ${GIT_DIR}
+    git clone git@github.com:minimization/feedback-pipeline-config.git
+  fi
+  if ! [ -d ${GIT_DIR}/feedback-pipeline-config ] ; then
+  	echo
+  	echo "You do not seem to have correct credentials for the git repo"
+  	echo "Exiting so you do no harm"
+  	echo
+  	exit 5
+  fi
+  cd ${GIT_DIR}/feedback-pipeline-config/configs
+  git pull
+  cp ${PACKAGELIST_DIR}/${VIEW}-buildroot-workload.yaml .
+  git add ${VIEW}-buildroot-workload.yaml
+  git commit -m "Update ${VIEW}-buildroot-workload $(date +%Y-%m-%d-%H:%M)"
+  git push
+fi # end - only update if not released repo
 
 exit 0
